requirement.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.8.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
boto3==1.34.139
botocore==1.34.139
build==1.2.1
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
clarifai==10.5.3
clarifai-grpc==10.6.0
click==8.1.7
cohere==5.5.8
colorama==0.4.6
coloredlogs==15.0.1
contextlib2==21.6.0
crewai==0.35.8
crewai-tools==0.4.6
dataclasses-json==0.6.7
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
docker==7.1.0
docstring_parser==0.16
docx2txt==0.8
email_validator==2.2.0
embedchain==0.1.113
fastapi==0.111.0
fastapi-cli==0.0.4
fastavro==1.9.4
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
google-api-core==2.19.1
google-auth==2.31.0
google-cloud-aiplatform==1.58.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
googletrans==4.0.0rc1
gptcache==0.1.43
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.7.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
inquirerpy==0.3.4
instructor==1.3.3
Jinja2==3.1.4
jiter==0.4.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
kubernetes==30.1.0
lancedb==0.5.7
langchain==0.1.20
langchain-cohere==0.1.5
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.7
langchain-text-splitters==0.0.2
langsmith==0.1.83
libretranslatepy==2.1.1
lxml==5.2.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
memzero==0.0.7
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.35.13
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.6
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parameterized==0.9.0
pfzy==0.3.4
pillow==10.4.0
pluggy==1.5.0
posthog==3.5.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
pulsar-client==3.5.0
py==1.11.0
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pylance==0.9.18
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pyright==1.1.370
pysbd==0.3.4
PySocks==1.7.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-rapidjson==1.18
pytube==15.0.0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
ratelimiter==1.2.0.post0
regex==2023.12.25
requests==2.32.3
requests-oauthlib==2.0.0
retry==0.9.2
rfc3986==1.5.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.2
schema==0.7.5
selenium==4.22.0
semver==3.0.2
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
starlette==0.37.2
sympy==1.12.1
tabulate==0.9.0
tenacity==8.4.2
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.4
translate==3.6.1
trio==0.26.0
trio-websocket==0.11.1
tritonclient==2.47.0
typer==0.12.3
types-requests==2.32.0.20240622
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
watchfiles==0.22.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2
